DONE - register, login, authentication

+++ backend/src/controllers/user.controller.js
@@ -1,7 +1,7 @@
 import userService from "../services/user.service.js";
-// import { JWT_SECRET } from "../constants/constants.js";
-// import HttpError from "../utils/HttpError.js";
-// import { extractUserIdFromToken } from "../utils/validation.utils.js";
+import { JWT_SECRET } from "../constants/constants.js";
+import HttpError from "../utils/HttpError.js";
+import { extractUserIdFromToken } from "../utils/validation.utils.js";
 
 const create = async (req, res, next) => {
   const { username, email, password } = req.body;
@@ -41,10 +41,10 @@ const update = async (req, res, next) => {
   const { username, email, password, isAdmin } = req.body;
 
   // UserId form token - AUTHENTICATION NEEDED
-  //   let userId = extractUserIdFromToken(req, JWT_SECRET);
-  //   if (id != userId) {
-  //     return next(new HttpError("Csak a saját profilodat módosíthatod", 401));
-  //   }
+  let userId = extractUserIdFromToken(req, JWT_SECRET);
+  if (id != userId) {
+    return next(new HttpError("Csak a saját profilodat módosíthatod", 401));
+  }
 
   try {
     const updatedUser = await userService.update(id, {
@@ -63,10 +63,10 @@ const destroy = async (req, res, next) => {
   const { id } = req.params;
 
   // UserId form token - AUTHENTICATION NEEDED
-  //   let userId = extractUserIdFromToken(req, JWT_SECRET);
-  //   if (id != userId) {
-  //     return next(new HttpError("Csak a saját profilodat törölheted", 401));
-  //   }
+  let userId = extractUserIdFromToken(req, JWT_SECRET);
+  if (id != userId) {
+    return next(new HttpError("Csak a saját profilodat törölheted", 401));
+  }
 
   try {
     const deletedUser = await userService.destroy(id);
